requirements.txt
Pillow
PyMuPDF
PySide6
nuitka
ordered-set